BasesAndPyQT/requirements.txt
pyyaml
jupyter
chardet
requests
pytest
pytest-cov
fabric3
tabulate
sqlalchemy
PyQt5
pycryptodome
sphinx
